Split off dial.rs
